funciones y test
