--- otro-index.js
@@ -1,8 +1,0 @@
-const nodejs = require('./nodejs.js');
-console.log(nodejs);
-
-/*console.log(nodejs.add(2, 3));
-console.log(nodejs.substract(10, 3));
-console.log(nodejs.multiply(5, 4));
-console.log(nodejs.divide(0, 5));
-console.log(nodejs.divide(20, 5));*/
