+++ package.json
@@ -5,6 +5,22 @@
 		"jest": "^24.7.1"
 	},
 	"scripts": {
-		"test": "jest"
-	}
+		"test": "jest --coverage --colors"
+	},
+	"description": "## Preámbulo",
+	"main": "index.js",
+	"directories": {
+		"test": "test"
+	},
+	"dependencies": {},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/xatzel/GDL002-md-links.git"
+	},
+	"author": "Xalex Ixbalamque Hernandez",
+	"license": "ISC",
+	"bugs": {
+		"url": "https://github.com/xatzel/GDL002-md-links/issues"
+	},
+	"homepage": "https://github.com/xatzel/GDL002-md-links#readme"
 }
