+++ test/md-links.spec.js
@@ -1,25 +1,15 @@
-/*const mdLinks = require('../');
-
-
-describe('mdLinks', () => {
-
-  it('should...', () => {
-    console.log('FIX ME!');
-  });
-
-})*/
-const { validatePath, absolutePath, readFileMd } = require('../src/mdLink.js');
+const { validatePath, absolutePath, readAllDocs } = require('../index.js');
 
 test('should be a file markdown', () => {
 	expect(validatePath('README.md')).toBe(true);
 });
 
 test('should be a absolute path ', () => {
-	expect(absolutePath('README.md')).toBe('/home/laboratoria-180/Escritorio/GDL002-md-links/README.md');
+	expect(absolutePath('README.md')).toBe('/Users/xary/Documents/laboratoria/GDL002-md-links/README.md');
 });
 
 test('should read a file', () => {
-	readFileMd('prueba.md').then((result) => {
+	readAllDocs('prueba.md').then((result) => {
 		expect(result).toBe('holis a todos');
 	});
 });
